use Strict language extension; simplify annotation
Instead of transforming the generated Core code to enforce CBV semantics, we now rely on the Strict language extension. This fixes a bug, where the plugin would report an internal error when checking an expression of the form `delay (let _ = adv d in ())`. The problem is that the let binding is erased when translating the typed AST to Core, hence the clock elaborator cannot infer the clock. Alternatively, we could implement the clock elaboration at the typed AST level instead of Core, but the Strict language extension is much simpler and properly enforces a CBV semantics (e.g. by not eliminating unused let bindings).

Since language extensions are applied per module, we also change the way we use annotations. Now, if the plugin is enabled in a module, all definitions in the module are considered Async Rattus code.